+++ src/main/java/nextstep/subway/section/ui/SectionController.java
@@ -1,5 +1,6 @@
 package nextstep.subway.section.ui;
 
+import nextstep.subway.section.dto.SectionRequest;
 import nextstep.subway.section.dto.SectionResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,7 +11,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class SectionController {
     @PostMapping("/lines/{id}/sections")
-    public ResponseEntity<SectionResponse> createSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
+    public ResponseEntity<SectionResponse> createSection(@PathVariable Long id,
+                                                         @RequestBody SectionRequest sectionRequest) {
         return null;
     }
 }
